pyproject.toml
[project]
name = "basic-task-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.178.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.29.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "ollama>=0.5.3",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
]
